# httpapi/openapi.yaml
openapi: "3.1.0"
info:
  title: Pit Officer REST API
  version: "1"
  description: >
    Control-plane REST API for the Pit Officer service. The public surface is
    mounted at `/api/v1`; a mirror at `/app/api/v1` is reserved for the operator
    dashboard and carries the same routes except the route manifest, which is
    published only on `/api/v1`. Control endpoints answer `application/json`;
    endpoints that hand back a file answer with that file's own media type.
    A malformed request body, or one that violates the schema or a parameter
    constraint, answers RFC 9457 `application/problem+json` - `400` for broken
    syntax, `422` for a constraint violation. Every other error answers the
    envelope `{"error":{"code":"...","message":"..."}}`.
    Money and size values
    are exact decimal strings - never IEEE 754 floats.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: /api/v1
    description: Public REST API
...
paths:
...
  /accounts:
    get:
      operationId: listAccounts
      summary: List all accounts
      tags: [accounts]
      parameters:
        - $ref: "#/components/parameters/CodeFilter"
        - $ref: "#/components/parameters/CodeMatch"
        - $ref: "#/components/parameters/StatusFilter"
        - $ref: "#/components/parameters/PositionCountMode"
        - $ref: "#/components/parameters/PositionCountMin"
        - $ref: "#/components/parameters/PositionCountMax"
        - name: pnlMode
          in: query
          schema:
            type: string
            enum: [all, eq, equal, equals, exact, neq, not_equal, not_equals, gt, greater_than, gte, lt, less_than, lte, between]
            default: all
          description: Compare the account P&L, independently of position P&L.
        - name: pnlMin
          in: query
          schema:
            type: string
          description: Exact decimal lower bound for account P&L.
        - name: pnlMax
          in: query
          schema:
            type: string
          description: Exact decimal upper bound for account P&L.
        - name: pnlCurrency
          in: query
          schema:
            type: string
          description: >-
            Required for P&L bounds or category flags. Only meaningful amounts
            in this resolved account currency are compared; no conversion is
            performed. Explicit category flags also admit non-numeric rows,
            which carry no amount to compare in any currency.
        - name: includeUnsetPnl
          in: query
          schema:
            type: boolean
            default: false
          description: Include unset account P&L alongside matching amounts.
        - name: includeHaltedPnl
          in: query
          schema:
            type: boolean
            default: false
          description: Include halted account P&L alongside matching amounts.
        - name: blockReason
          in: query
          schema:
            type: string
          description: >-
            Search the effective block reason - the account's own reason, or its
            group's when the block comes from the group.
        - $ref: "#/components/parameters/BlockReasonMatch"
        - name: group
          in: query
          schema:
            type: string
          description: Group code. Present with an empty value selects ungrouped accounts.
        - name: sort
          in: query
          schema:
            type: string
            enum: [code, title, group, status, blockReason, positionCount]
          description: >-
            Account list sort key. Unknown keys are rejected. `blockReason`
            orders by the effective reason, the same value the filter searches.
        - $ref: "#/components/parameters/SortOrder"
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/OffsetParam"
        - $ref: "#/components/parameters/PageParam"
      responses:
        "200":
          description: Account list.
          content:
            application/json:
              schema:
                type: object
                required: [accounts, total]
                properties:
                  accounts:
                    type: array
                    items:
                      $ref: "#/components/schemas/Account"
                  total:
                    type: integer
                    description: Total rows matching the filters before paging.
        "400":
          $ref: "#/components/responses/MalformedRequestError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      operationId: createAccount
      summary: Create an account
      tags: [accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [code]
              properties:
                code:
                  type: string
                  description: Account public code.
                title:
                  type: string
                  maxLength: 256
                  description: Mutable display title for the account.
                currency:
                  type: string
                  description: Optional account currency asset code. Empty or omitted inherits from group/default tiers.
      responses:
        "201":
          description: Account created.
          content:
            application/json:
              schema:
                type: object
                required: [account]
                properties:
                  account:
                    $ref: "#/components/schemas/Account"
        "400":
          $ref: "#/components/responses/MalformedRequestError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "500":
          $ref: "#/components/responses/InternalError"
  /accounts/{code}:
    parameters:
      - $ref: "#/components/parameters/AccountCode"
    get:
      operationId: getAccount
      summary: Get account detail and its typed limit barriers
      tags: [accounts]
      responses:
        "200":
          description: Account and its limits.
          content:
            application/json:
              schema:
                type: object
                required: [account, limits]
                properties:
                  account:
                    $ref: "#/components/schemas/Account"
                  limits:
                    $ref: "#/components/schemas/AccountLimits"
        "400":
          $ref: "#/components/responses/MalformedRequestError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"
    put:
      operationId: updateAccount
      summary: Update account code and title
      tags: [accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [code, title]
              properties:
                code:
                  type: string
                  description: Replacement account public code.
                title:
                  type: string
                  maxLength: 256
                  description: Replacement display title.
      responses:
        "200":
          description: Account after update.
          content:
            application/json:
              schema:
                type: object
                required: [account]
                properties:
                  account:
                    $ref: "#/components/schemas/Account"
        "400":
          $ref: "#/components/responses/MalformedRequestError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      operationId: deleteAccount
      summary: Delete an account
      tags: [accounts]
      parameters:
        - $ref: "#/components/parameters/ForceCascade"
      responses:
        "204":
          description: Account deleted.
        "400":
          $ref: "#/components/responses/MalformedRequestError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "409":
          $ref: "#/components/responses/HasDependentsError"
        "503":
          $ref: "#/components/responses/EngineRestartingError"
        "500":
          $ref: "#/components/responses/InternalError"
  /accounts/{code}/block:
    parameters:
      - $ref: "#/components/parameters/AccountCode"
    post:
      operationId: blockAccount
      summary: Block an account
      tags: [accounts]
      parameters:
        - $ref: "#/components/parameters/MissingAccount"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [reason]
              properties:
                reason:
                  type: string
                  description: Human-readable reason for the block.
      responses:
        "200":
          description: Account after blocking.
          content:
            application/json:
              schema:
                type: object
                required: [account]
                properties:
                  account:
                    $ref: "#/components/schemas/Account"
        "400":
          $ref: "#/components/responses/MalformedRequestError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/AccountMissingOrNotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"
  /accounts/{code}/unblock:
    parameters:
      - $ref: "#/components/parameters/AccountCode"
    post:
      operationId: unblockAccount
      summary: Unblock an account
      tags: [accounts]
      parameters:
        - $ref: "#/components/parameters/MissingAccount"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
      responses:
        "200":
          description: Account after unblocking.
          content:
            application/json:
              schema:
                type: object
                required: [account]
                properties:
                  account:
                    $ref: "#/components/schemas/Account"
        "400":
          $ref: "#/components/responses/MalformedRequestError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/AccountMissingOrNotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"
  /accounts/{code}/group:
    parameters:
      - $ref: "#/components/parameters/AccountCode"
    put:
      operationId: setAccountGroup
      summary: Set (or clear) account group membership
      description: >
        Assigns the account to a group. Pass an empty string to clear group
        membership. The move returns 409 `currency_change_blocked` when it would
        alter the account's effective currency while the account has any
        non-zero balance, position, cost-basis, or P&L value. An account with a
        currency of its own resolves through no group tier, so its membership
        moves freely.
      tags: [accounts]
      parameters:
        - $ref: "#/components/parameters/MissingAccount"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [group]
              properties:
                group:
                  type: string
                  description: Group code. Empty string clears membership.
      responses:
        "200":
          description: Account after update.
          content:
            application/json:
              schema:
                type: object
                required: [account]
                properties:
                  account:
                    $ref: "#/components/schemas/Account"
        "400":
          $ref: "#/components/responses/MalformedRequestError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/AccountMissingOrNotFoundError"
        "409":
          $ref: "#/components/responses/CurrencyChangeBlockedError"
        "500":
          $ref: "#/components/responses/InternalError"
  /accounts/{code}/currency:
    parameters:
      - $ref: "#/components/parameters/AccountCode"
    put:
      operationId: setAccountCurrency
      summary: Set or clear account currency
      description: >
        Sets the account-level currency asset code. Pass an empty string to clear
        the account tier and inherit from group/default tiers. Changing the
        effective currency returns 409 when the account has any non-zero balance,
        position, cost-basis, or P&L value.
      tags: [accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [currency]
              properties:
                currency:
                  type: string
                  description: Asset code to use as the account currency, or empty to clear it.
      responses:
        "200":
          description: Account after currency update.
          content:
            application/json:
              schema:
                type: object
                required: [account]
                properties:
                  account:
                    $ref: "#/components/schemas/Account"
        "400":
          $ref: "#/components/responses/MalformedRequestError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "409":
          $ref: "#/components/responses/CurrencyChangeBlockedError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /accounts/{code}/notes:
    parameters:
      - $ref: "#/components/parameters/AccountCode"
    put:
      operationId: setAccountNotes
      summary: Set operator notes on an account
      tags: [accounts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [notes]
              properties:
                notes:
                  type: string
                  description: Free-form operator notes stored on the account.
      responses:
        "200":
          description: Account after update.
          content:
            application/json:
              schema:
                type: object
                required: [account]
                properties:
                  account:
                    $ref: "#/components/schemas/Account"
        "400":
          $ref: "#/components/responses/MalformedRequestError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /groups:
    get:
      operationId: listGroups
      summary: List all account groups
      tags: [groups]
      parameters:
        - $ref: "#/components/parameters/CodeFilter"
        - $ref: "#/components/parameters/CodeMatch"
        - $ref: "#/components/parameters/StatusFilter"
        - $ref: "#/components/parameters/PositionCountMode"
        - $ref: "#/components/parameters/PositionCountMin"
        - $ref: "#/components/parameters/PositionCountMax"
        - name: accountCount
          in: query
          schema:
            type: string
            enum: [all, has, none]
            default: all
          description: Filter by whether the group has member accounts.
        - $ref: "#/components/parameters/AccountCountMode"
        - $ref: "#/components/parameters/AccountCountMin"
        - $ref: "#/components/parameters/AccountCountMax"
        - name: notes
          in: query
          schema:
            type: string
          description: Search group notes.
        - $ref: "#/components/parameters/NotesMatch"
        - name: blockReason
          in: query
          schema:
            type: string
          description: Search group block reason.
        - $ref: "#/components/parameters/BlockReasonMatch"
        - name: sort
          in: query
          schema:
            type: string
            enum: [code, title, accountCount, positionCount, status, notes, blockReason]
          description: Group list sort key. Unknown keys are rejected.
        - $ref: "#/components/parameters/SortOrder"
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/OffsetParam"
        - $ref: "#/components/parameters/PageParam"
      responses:
        "200":
          description: Group list.
          content:
            application/json:
              schema:
                type: object
                required: [groups, total]
                properties:
                  groups:
                    type: array
                    items:
                      $ref: "#/components/schemas/Group"
                  total:
                    type: integer
                    description: >
                      Total real groups matching the filters before paging. The
                      synthetic default group is not counted.
        "400":
          $ref: "#/components/responses/MalformedRequestError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      operationId: createGroup
      summary: Create an account group
      tags: [groups]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [code]
              properties:
                code:
                  type: string
                  description: >-
                    Group public code. `-` is rejected with `validation`: it is
                    reserved for addressing the realm default group, so a group
                    carrying it could never be reached through any group
                    operation.
                title:
                  type: string
                  maxLength: 256
                  description: Mutable display title for the group.
                notes:
                  type: string
                  maxLength: 4096
                  description: Free-form operator notes stored on the group.
                currency:
                  type: string
                  description: Optional group currency asset code. Empty or omitted leaves the group tier unset.
      responses:
        "201":
          description: Group created.
          content:
            application/json:
              schema:
                type: object
                required: [group]
                properties:
                  group:
                    $ref: "#/components/schemas/Group"
        "400":
          $ref: "#/components/responses/MalformedRequestError"
        "422":
          $ref: "#/components/responses/ValidationError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
...
  parameters:
    AccountCode:
      name: code
      in: path
      required: true
      description: URL-encoded account public code.
      schema:
        type: string
...
    CodeFilter:
      name: code
      in: query
      description: >
        Code search text. A plain value is matched as a substring. `*` is a
        friendly wildcard whose fragments must appear in order.
      schema:
        type: string
    CodeMatch:
      name: codeMatch
      in: query
      description: Anchoring mode applied to the code filter.
      schema:
        type: string
        enum: [contains, starts_with, ends_with, exact]
        default: contains
...
    BlockReasonMatch:
      name: blockReasonMatch
      in: query
      description: Anchoring mode applied to the block-reason filter.
      schema:
        type: string
        enum: [contains, starts_with, ends_with, exact]
        default: contains
    StatusFilter:
      name: status
      in: query
      description: >-
        Filter by blocked state. On the account list this is the effective
        state: `blocked` also selects accounts blocked only through their
        group, and `active` excludes them.
      schema:
        type: string
        enum: [all, active, blocked]
        default: all
    PositionCountMode:
      name: positionCountMode
      in: query
      description: Position count comparison mode.
      schema:
        type: string
        enum: [all, eq, equal, equals, exact, neq, not_equal, not_equals, gt, greater_than, gte, lt, less_than, lte, between]
        default: all
    PositionCountMin:
      name: positionCountMin
      in: query
      description: Lower position-count bound. Required for greater_than and between.
      schema:
        type: integer
        minimum: 0
    PositionCountMax:
      name: positionCountMax
      in: query
      description: Upper position-count bound. Required for less_than and between.
      schema:
        type: integer
        minimum: 0
...
    SortOrder:
      name: order
      in: query
      description: Sort direction.
      schema:
        type: string
        enum: [asc, desc]
        default: asc
    ForceCascade:
      name: force
      in: query
      description: Allow application-approved cascade deletion of dependent rows.
      schema:
        type: boolean
        default: false
...
    LimitParam:
      name: limit
      in: query
      description: Maximum rows to return (default 100, capped at 1000).
      schema:
        type: integer
        default: 100
        maximum: 1000
    OffsetParam:
      name: offset
      in: query
      description: Zero-based row offset for server-side pagination.
      schema:
        type: integer
        default: 0
        minimum: 0
    PageParam:
      name: page
      in: query
      description: >
        One-based page number for server-side pagination. When present, the
        backend computes offset as `(page - 1) * limit`.
      schema:
        type: integer
        minimum: 1

  # ---------------------------------------------------------------------------
  # Responses
  # ---------------------------------------------------------------------------
  responses:
    MalformedRequestError:
      description: Malformed HTTP request or JSON syntax.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    ValidationError:
      description: Syntactically valid request failed schema or parameter validation.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
    NotFoundError:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ConflictError:
      description: >-
        Resource conflict, including an existing code, protected runtime state,
        or a configured dependency that prevents the mutation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    EngineRestartingError:
      description: Engine is being rebuilt; mutating requests are rejected.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    InternalError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  # ---------------------------------------------------------------------------
  # Schemas
  # ---------------------------------------------------------------------------
  schemas:
...
    AccountLimits:
      type: object
      required: [rateLimits, orderSizeLimits, spotFundsPnlBoundsLimits]
      properties:
        rateLimits:
          type: array
          description: Rate-limit barriers (always a non-null array).
          items:
            $ref: "#/components/schemas/RateLimit"
        orderSizeLimits:
          type: array
          description: Order-size barriers (always a non-null array).
          items:
            $ref: "#/components/schemas/OrderSizeLimit"
        spotFundsPnlBoundsLimits:
          type: array
          description: SpotFunds P&L-bounds barriers (always a non-null array).
          items:
            $ref: "#/components/schemas/SpotFundsPnlBoundsLimit"
...